Fix hotseat icon translation issue
The current behavior that adjusts hotseat icons to make room for the
bubble bar uses the index of the hotseat icon to calculate translation x.

The problem is that if icons are reordered, the index of the icon may be
different than the actual cell number it belongs to.

This change updates the logic to use the cell rather than the index.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 345686425
Test: manual - on tablet in portrait
       - show bubble bar
       - observe hotseat icons adjust
       - drag an icon from the home screen into hotseat
       - observe icons are shown correctly
       - rotate the device to landscape and back to portrait
Change-Id: I8628892234d1b31205b93ba3089c6dbf2820ea04