Animate the dot with the bubble bar
The previous behavior relied on a BEHIND_STACK suppression flag to
determine whether the dot should be drawn or not. setBehindStack was
called as part of updating bubble layout properties in BubbleBarView,
which happens at each animation frame of the bubble bar collapsing
and expanding.

Instead of using the suppression flag, we now directly set the dot
scale based on the progress of the collapse and expand animation.

This change also sets the badge visibility directly from BubbleBarView
and cleans up the BEHIND_STACK suppression flag.

Flag: com.android.wm.shell.enable_bubble_bar
Fixes: 349822532
Test: manual
       - have some bubbles and expand bar
       - trigger a new bubble X
       - observe bubble added with notification dot
       - tap on the selected bubble to collapse
       - observe the notification dot animates for bubble X
       - expand the bar
       - observe notification dot animates for bubble X
       - tap on bubble X
       - observe notification dot animates out
       - tap on a different bubble
       - observe notification dot does not reappear
Change-Id: I6efe0dfa8b4f73064cee8b45556ff0928f6f7c08