Don't animate the bubble that was dismissed by drag
This change stores the bubble that was dismissed by drag temporarily
in BubbleBarView so that when we remove the bubble, we can avoid
updating it as part of the animation.

This is step 1 in this change. Next I'll look into removing the
bubble immediately after dismissing it from the menu, and lastly
update sysui to not send the removal event back to launcher.

Flag: com.android.wm.shell.enable_bubble_bar
Bug: 349826879
Test: manual
       - have bubbles in the bubble bar and expand it
       - drag a bubble to dismiss it
       - observe the bubble does not reappear in the bar during
         the animation
Change-Id: I6628ce779db4aac8d82edd4ecf2a25dedf8feef9