Prevent Taskbar edu dismissal from touching outside of the window.
In an effort to encourage users to look at Taskbar edu more thoroughly,
this change makes it so users have to press the 'Done' button explicitly
to close the taskbar edu window. This change also adds the ability for
it to be dismissed by the back gesture. It only applies to the features
edu, and not other taskbar educations.

Fix: 349612575
Test: Manually run all versions of taskbar edu and ensure that the edu only goes away with back gesture, swipe up or clicking on the DONE button
Flag: EXEMPT bugfix
Change-Id: I9c456b9efddc6de5f292d6a14b1ce9daee7efaa3