Replace Lottie animations for All Set page at the end of SUW
This change updates the background lottie animations that play on the
final page of Setup Wizard, the All Set page. They needed to be updated
since their colors would sometimes get corrupted, which made it
difficult to read the text on screen.

Fix: 318455805
Test: Manually run Setup wizard with the maximum font size and display
size and ensure the animations display correctly for different devices.
Flag: EXEMPT bugfix

Change-Id: I8f30c8e547aa1d704b5be7ba355719643749e3e4