Re-enable FREEFORM task shortcut when DW is unsupported
The freeform shortcut was previously changed to only show if the desktop
windowing flag was disabled, which is incorrect since it's not a
reliable indicator about whether desktop windowing is supported or not.
When the flag is enabled as the feature progresses, it causes the button
to be hidden on devices that support freeform windows but don't support
desktop windowing, such as phones or tablets when freeform is enabled in
developer options.

To properly check that DW is not enabled, this CL changes the flag check
to DesktopModeStatus#canEnterDesktopMode.

Bug: 353740331
Test: enable freeform in dev options, disable desktop windowing - check
"freeform" button is available through long-press of the app icon in
overview/recents.
Flag: EXEMPT bugfix

Change-Id: I0ea5073e5570838acf49a27f54546fd97e6739cd