Pass empty srcRectHint if overlay is present
Launcher should pass an empty srcRectHint to Shell
via stopSwipePipToHome if overlay is present, so
that Shell know that srcRectHint is being "simulated"
for the enter PiP animation with an app icon overlay.

Shell will recalculate the simulated src-rect based on
appBounds and aspect ratio of the destination bounds.

Bug: 352773027
Flag: EXEMPT bugfix
Test: swipe PiP to home from Maps with fixed rotation (90 and 270)
Change-Id: I5f89c5af9d6b8a66725e346d09e7b3d84044bb4c