Fix stuck focus outline after returning to home
Sometimes SmartspaceViewContainer is not the LastFocusedItem in ItemFocusIndicatorHelper when triggering changeFocus, it would skip the hiding outline animation which causes the stuck focus outline.
So the solution is to add a checking in "changeFocus" that if the LastFocusedItem is not the view calling changeFocus and it does not have focus now, it means no need to add or remove outline and could just return.

Bug: 349019294
Test: after going to notification panel, touch/click could make the outline disappears. https://drive.google.com/file/d/1aXQ38CEqvzJVj7gBM-KNTCJSpzfwYZsn/view?usp=sharing
Test: after going home from app in overview, touch/click make the outline disappears. https://drive.google.com/file/d/1DgYzAgIJqAHJ_DuYjswlCUBOiohprJze/view?usp=sharing
Flag: EXEMPT bugfix
Change-Id: I4613ba2fe92ec609055e48dc0b5ff53c2037a158