Show Taskbar when Glanceable Hub is showing.
This changelist ensures that the Taskbar is not hidden when the
Glanceable Hub is showing. Logic for hiding the Taskbar when dreaming
was suppressing it from showing when Glanceable Hub was present. The
surrounding condition has been updated to now not trigger is Glanceable
Hub is showing.

Test: manual - verified Taskbar was properly displayed and hidden
      between glanceable hub and dreaming.
Flag: EXEMPT bugfix
Fixes: 353803584
Change-Id: I58a61617baff65041757ae0df96687a071120880