Only handle taps on bubble bar when it was collapsed/stashed on down
When bubbles are expanded and you tap on the currently opened
bubble, the stack collapses. In the input consumer code, we also
get that tap & we handle it because at the "up" event at that time,
we're collapsed (or stashed in some cases). To fix this, only handle
the "up" event for the bar being collapsed / stashed, if it was in
that state when the gesture started (i.e. on down)

Flag: com.android.wm.shell.enable_bubble_bar
Bug: 362165243
Test: manual - expand bubble bar in home / overview
             - tap on the currently expanded bubble
             => observe the animation is smooth
             - expand bubble bar in app
             - tap on currently expanded bubble
             => observe the animation is smooth
Change-Id: I1861af15c88bc0cbf2514faac7464de2f85aad1e