Fix an issue where the bubble bar was visible after theme change
Don't set the alpha to 1 in the drag reset method unless the bar
is actually expanded.

When theme changes, everything gets recreated... so the location of
the bubble bar gets set which triggers drag reset which would make
the bar visible when it shouldn't be.

Flag: com.android.wm.shell.enable_bubble_bar
Test: manual - have bubbles in app (stashed), change theme, observe
               that bubbles are not visible
             - test dragging bubbles from left  to right via expanded
               view, individual bubble, and collapsed stack
Bug: 358304168 (not exact bug but noted in the video of comment#13)
Change-Id: I291cdb206532efe386684e36a78b9d66018f3cb1